+++ ocaml-lsp-server/src/merlin_config.mli
@@ -4,6 +4,8 @@ open Import
 
 type t
 
+val should_read_dot_merlin : bool ref
+
 val config : t -> Mconfig.t Fiber.t
 
 val destroy : t -> unit Fiber.t