+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -1094,6 +1094,7 @@ let start () =
              Server.notification server (Server_notification.ShowMessage log)))
     ]
 
-let run () =
+let run ~read_dot_merlin () =
+  Merlin_config.should_read_dot_merlin := read_dot_merlin;
   Unix.putenv "__MERLIN_MASTER_PID" (string_of_int (Unix.getpid ()));
   Lev_fiber.run (Lev.Loop.default ()) ~f:start