+++ ocaml-lsp-server/src/ocaml_lsp_server.mli
@@ -1,4 +1,4 @@
-val run : unit -> unit
+val run : read_dot_merlin:bool -> unit -> unit
 
 module Diagnostics = Diagnostics
 module Version = Version